moved tab bar helpers to separate file and added comments
